Fix distribution for windows

+++ tasks/morozov_n_block_gauss_filter/tests/functional/main.cpp
@@ -81,14 +81,14 @@ class MorozovNBlockGaussFilterFuncTestsProcesses : public ppc::util::BaseRunFunc
     }
     std::vector<uint8_t> img(width * height * 3, 0);
     std::mt19937 gen(seed);
-    std::uniform_int_distribution<uint8_t> dis(0, 255);
+    std::uniform_int_distribution<uint32_t> dis(0, 255);
 
     for (int i = 0; i < height; i++) {
       for (int j = 0; j < width; j++) {
         int pixel_idx = 3 * ((i * width) + j);
-        img[pixel_idx + 0] = dis(gen);  // R
-        img[pixel_idx + 1] = dis(gen);  // G
-        img[pixel_idx + 2] = dis(gen);  // B
+        img[pixel_idx + 0] = static_cast<uint8_t>(dis(gen));  // R
+        img[pixel_idx + 1] = static_cast<uint8_t>(dis(gen));  // G
+        img[pixel_idx + 2] = static_cast<uint8_t>(dis(gen));  // B
       }
     }
     return std::make_tuple(img, width, height);
@@ -117,61 +117,6 @@ class MorozovNBlockGaussFilterFuncTestsProcesses : public ppc::util::BaseRunFunc
     std::string abs_path = ppc::util::GetAbsoluteTaskPath(PPC_ID_morozov_n_block_gauss_filter, relative_path);
     return ReadImageFromFile(abs_path);
   }
-
-  // Save image to file using stb_image_write
-  // static bool SaveImageToFile(const std::vector<uint8_t> &img_data, int width, int height,
-  //                              const std::string &filename) {
-  //   if (img_data.size() < static_cast<size_t>(width * height * 3)) {
-  //     return false;
-  //   }
-  //   int result = stbi_write_png(filename.c_str(), width, height, 3, img_data.data(), width * 3);
-  //   return result != 0;
-  // }
-
-  // // Save result images (correct and actual) for comparison
-  // void SaveResultImages(const OutType &output_data) {
-  //   try {
-  //     // Get test parameters for naming
-  //     TestType params = std::get<static_cast<std::size_t>(ppc::util::GTestParamIndex::kTestParams)>(GetParam());
-  //     std::string test_name = std::get<static_cast<std::size_t>(ppc::util::GTestParamIndex::kNameTest)>(GetParam());
-  //     const std::string &test_type = std::get<0>(params);
-  //     int width = std::get<1>(params);
-  //     int height = std::get<2>(params);
-
-  //     // Get actual dimensions from input
-  //     int img_width = std::get<1>(input_data_);
-  //     int img_height = std::get<2>(input_data_);
-
-  //     // Create output directory
-  //     std::filesystem::path output_dir;
-  //     const char *tmpdir = std::getenv("PPC_TEST_TMPDIR");
-  //     if (tmpdir != nullptr) {
-  //       output_dir = std::filesystem::path(tmpdir) / "filter_results";
-  //     } else {
-  //       output_dir = std::filesystem::path("test_output") / "filter_results";
-  //     }
-  //     std::filesystem::create_directories(output_dir);
-
-  //     // Generate unique filename based on test name and parameters
-  //     std::string base_name = test_name;
-  //     if (test_type == "gen") {
-  //       base_name += "_gen_" + std::to_string(width) + "x" + std::to_string(height);
-  //     } else {
-  //       base_name += "_" + test_type;
-  //     }
-
-  //     // Save correct result
-  //     std::string correct_filename = (output_dir / (base_name + "_correct.png")).string();
-  //     SaveImageToFile(correct_data_, img_width, img_height, correct_filename);
-
-  //     // Save actual result
-  //     std::string actual_filename = (output_dir / (base_name + "_actual.png")).string();
-  //     SaveImageToFile(output_data, img_width, img_height, actual_filename);
-  //   } catch (const std::exception &e) {
-  //     // Silently fail - image saving is for debugging, shouldn't break tests
-  //     (void)e;
-  //   }
-  // }
 };
 
 namespace {
@@ -180,8 +125,6 @@ TEST_P(MorozovNBlockGaussFilterFuncTestsProcesses, MatmulFromPic) {
   ExecuteTest(GetParam());
 }
 
-// Test parameters: (test_type, width, height)
-// test_type can be "gen" for generated images or image filename without extension for file-based tests
 const std::array<TestType, 5> kTestParam = {
     std::make_tuple("img_1", 0, 0),    // Read from file img_1.jpg
     std::make_tuple("gen", 50, 50),    // Generated 50x50 image
